Update Recursive Factorial.cpp

+++ Day7/Recursive Factorial.cpp
@@ -1,6 +1,6 @@
 #include<iostream>
 using namespace std;
-long long factorial(int n){
+int factorial(int n){
     if(n==0 || n==1) return 1;
     return n*factorial(n-1);
 }
